chore: restructure files to have less folders
