+++ app/layout.tsx
@@ -1,13 +1,11 @@
-import { Geist, Geist_Mono } from "next/font/google";
-import { Providers } from "@/providers/providers";
-import { siteConfig } from "@/config/site-config";
-import { Toast } from "@/components/ui/toast";
-import type { Metadata } from "next";
-import "./globals.css";
-
 import MailComposeModal from "@/components/mail/mail-compose-modal";
-import { NuqsAdapter } from "nuqs/adapters/next/app";
+import { Geist, Geist_Mono } from "next/font/google";
+import { siteConfig } from "@/lib/site-config";
+import { Toast } from "@/components/ui/toast";
+import { Providers } from "@/lib/providers";
+import { cn } from "@/lib/utils";
 import { Suspense } from "react";
+import "./globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -19,7 +17,7 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-export const metadata: Metadata = siteConfig;
+export const metadata = siteConfig;
 
 export default function RootLayout({
   children,
@@ -28,15 +26,13 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" suppressHydrationWarning>
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+      <body className={cn(geistSans.variable, geistMono.variable, "antialiased")}>
         <Providers attribute="class" defaultTheme="dark" enableSystem disableTransitionOnChange>
-          <NuqsAdapter>
-            <Suspense>
-              <MailComposeModal />
-            </Suspense>
-            {children}
-            <Toast />
-          </NuqsAdapter>
+          <Suspense>
+            <MailComposeModal />
+          </Suspense>
+          {children}
+          <Toast />
         </Providers>
       </body>
     </html>
