+++ components/context/sidebar-context.tsx
@@ -9,8 +9,7 @@ import {
 } from "@/lib/constants";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { TooltipProvider } from "../ui/tooltip";
-import { getCookie } from "@/utils/cookies";
-import { cn } from "@/lib/utils";
+import { cn, getCookie } from "@/lib/utils";
 import * as React from "react";
 
 export type SidebarContext = {
