+++ lib/constants.ts
@@ -5,6 +5,7 @@ export const SIDEBAR_WIDTH_MOBILE = "12rem";
 export const SIDEBAR_WIDTH_ICON = "3rem";
 export const SIDEBAR_KEYBOARD_SHORTCUT = "b";
 export const BASE_URL = process.env.NEXT_PUBLIC_APP_URL;
+export const MAX_URL_LENGTH = 2000;
 export const ALLOWED_HTML_TAGS = [
   "p",
   "br",
