+++ lib/utils.ts
@@ -1,12 +1,10 @@
+import { format, isToday, isThisMonth, differenceInCalendarMonths } from "date-fns";
+import { MAX_URL_LENGTH } from "./constants";
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 import LZString from "lz-string";
 
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
-}
-
-const MAX_URL_LENGTH = 2000;
+export const cn = (...inputs: ClassValue[]) => twMerge(clsx(inputs));
 
 export const compressText = (text: string): string => {
   const compressed = LZString.compressToEncodedURIComponent(text);
@@ -15,4 +13,18 @@ export const compressText = (text: string): string => {
 
 export const decompressText = (compressed: string): string => {
   return LZString.decompressFromEncodedURIComponent(compressed) || "";
+};
+
+export const getCookie = (key: string): string | null => {
+  const cookies = Object.fromEntries(
+    document.cookie.split("; ").map((v) => v.split(/=(.*)/s).map(decodeURIComponent)),
+  );
+  return cookies?.[key] ?? null;
+};
+
+export const formatDate = (date: string) => {
+  if (isToday(date)) return format(date, "h:mm a");
+  if (isThisMonth(date) || differenceInCalendarMonths(new Date(), date) === 1)
+    return format(date, "MMM dd");
+  return format(date, "MM/dd/yy");
 };
