--- providers/theme-provider.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import { ThemeProvider as NextThemesProvider } from "next-themes";
-import * as React from "react";
-
-export function ThemeProvider({
-  children,
-  ...props
-}: React.ComponentProps<typeof NextThemesProvider>) {
-  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
-}
