--- utils/cookies.ts
@@ -1,6 +1,0 @@
-export const getCookie = (key: string): string | null => {
-  const cookies = Object.fromEntries(
-    document.cookie.split("; ").map((v) => v.split(/=(.*)/s).map(decodeURIComponent)),
-  );
-  return cookies?.[key] ?? null;
-};
